3DDonut/requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
pygame>=2.1.0 